[#21090] Removed the separate class and refactored the code according to the comments
